requirements.txt
agno>=2.2.10
streamlit>=1.40.2
openai>=1.102.0
requests
firecrawl-py>=4.6.0
elevenlabs>=1.0.0